requirements.txt
python-whois==0.8.0
dnspython==2.6.1
rich==13.7.1

streamlit==1.35.0
streamlit==1.35.0
